# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HonMoon API",
    "description": "HonMoon API 문서",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/auth/signup/email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "이메일 회원가입 링크 전송",
        "description": "입력된 이메일로 회원가입 확인 링크를 전송합니다.",
        "operationId": "signupByEmail",
        "parameters": [
          {
            "name": "redirectUrl",
            "in": "query",
            "description": "인증 성공 후 리다이렉트할 프론트엔드 URL",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "https://honmoon.site/auth/success"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyEmailMagicLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password/set": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "비밀번호 설정/변경",
        "description": "세션 사용자에 대해 비밀번호를 설정하거나 변경합니다. (8자 이상)",
        "operationId": "setPassword",
        "parameters": [
          {
            "name": "principal",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyLogoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "로그아웃",
        "description": "서버 DB 세션을 무효화합니다.",
        "operationId": "logout",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyLogoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/email/password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "이메일 ID/비밀번호 로그인",
        "description": "이메일과 비밀번호로 로그인하여 세션 토큰을 발급합니다.",
        "operationId": "loginWithEmailPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailPasswordLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyAuthLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/email/by-user": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "이메일 로그인 링크 전송(사용자 ID)",
        "description": "프론트에서 이메일을 입력하지 않고 사용자 ID로 이메일을 조회하여 매직 링크를 전송합니다.",
        "operationId": "loginByEmailByUser",
        "parameters": [
          {
            "name": "redirectUrl",
            "in": "query",
            "description": "인증 성공 후 리다이렉트할 프론트엔드 URL",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "https://honmoon.site/auth/success"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginByUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyEmailMagicLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Google OAuth 코드 교환",
        "description": "프론트 콜백에서 받은 code/state를 서버에 전달하여 세션 토큰을 발급받습니다.",
        "operationId": "exchangeGoogleCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleCodeExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyAuthLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "이메일 매직 토큰 교환",
        "description": "매직 토큰을 검증하고 세션 토큰을 JSON으로 반환합니다.",
        "operationId": "exchangeEmailMagicToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailMagicTokenExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyEmailCallbackResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "현재 사용자 프로필",
        "description": "Basic 인증 또는 Bearer 세션 토큰으로 호출 시 현재 인증된 프로필을 반환합니다.",
        "operationId": "me",
        "parameters": [
          {
            "name": "principal",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/url": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Google OAuth 인증 URL 생성",
        "description": "프론트는 이 URL로 리다이렉트합니다. state는 위변조 방지를 위한 서명 값을 포함합니다.",
        "operationId": "getGoogleAuthUrl",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "description": "스코프 목록 공백 구분",
            "required": false,
            "schema": {
              "type": "string",
              "default": "profile email openid"
            },
            "example": "openid email profile"
          },
          {
            "name": "redirectAfter",
            "in": "query",
            "description": "인증 성공 후 프론트에서 이동할 경로",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "/"
          },
          {
            "name": "frontendCallbackUrl",
            "in": "query",
            "description": "프론트 콜백 URL (코드/상태를 넘길 경로)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "https://honmoon.site/auth/google/callback"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUrlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Google OAuth 콜백",
        "description": "구글에서 전달한 code/state로 토큰을 교환하고 사용자 정보를 조회한 뒤, 서버 DB 세션 토큰을 발급합니다.",
        "operationId": "googleCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "구글이 전달하는 인증 코드",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "요청 시 생성된 state",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "state 검증 실패 또는 파라미터 오류"
          }
        }
      }
    },
    "/api/auth/email/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "이메일 매직 링크 콜백",
        "description": "매직 토큰을 검증합니다.",
        "operationId": "magicLinkCallback",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "purpose",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirectUrl",
            "in": "query",
            "description": "인증 성공 후 리다이렉트할 프론트엔드 URL",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserPrincipal": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "provider",
          "roles",
          "subject"
        ]
      },
...
      "EmailSignUpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
      "EmailMagicLinkResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "magicLink": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "email",
          "expiresAt",
          "magicLink"
        ]
      },
      "ResponseBodyEmailMagicLinkResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EmailMagicLinkResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "SetPasswordRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "ResponseBodyLogoutResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LogoutResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "EmailPasswordLoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthLoginResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "google": {
            "$ref": "#/components/schemas/GoogleUserInfo"
          },
          "googleTokens": {
            "$ref": "#/components/schemas/GoogleTokenResponse"
          },
          "appSessionToken": {
            "type": "string"
          },
          "jwt": {
            "$ref": "#/components/schemas/JwtTokenResponse"
          }
        },
        "required": [
          "provider"
        ]
      },
      "GoogleTokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "idToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresInSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "scope": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "expiresInSeconds",
          "tokenType"
        ]
      },
      "GoogleUserInfo": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "givenName": {
            "type": "string"
          },
          "familyName": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          }
        },
        "required": [
          "sub"
        ]
      },
      "JwtTokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "tokenType": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt",
          "tokenType"
        ]
      },
      "ResponseBodyAuthLoginResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthLoginResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "EmailLoginByUserRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "userId"
        ]
      },
      "GoogleCodeExchangeRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "state"
        ]
      },
      "EmailMagicTokenExchangeRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "EmailCallbackResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "isValid": {
            "type": "boolean"
          },
          "appSessionToken": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "isValid"
        ]
      },
      "ResponseBodyEmailCallbackResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EmailCallbackResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "sub"
        ]
      },
      "AuthUrlResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "authorizationUrl": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "authorizationUrl",
          "provider",
          "state"
        ]
      }
    }
  }
}